requirements.txt
cryptography
pycryptodome
smtplib
email

